borrar carpeta fix

+++ controladores/carpetas/eliminarCarpeta.js
@@ -9,27 +9,31 @@ const eliminarCarpeta = async (req, res, next) => {
     const idUsuario = req.idPropietario;
 
     const borrar = await mostrarFicheros(idCarpeta, idUsuario);
-    const archivosBorrar = [...borrar[1]];
 
-    if (archivosBorrar.length > 0 || archivosBorrar != undefined) {
-      //recorro el array de ficheros pertenecientes a la carpeta a borrar
-      for (const archivo of archivosBorrar) {
-        //borro el fichero en la base de datos
-        let ficheroBorrado = borrarFichero(idUsuario, archivo.id_archivo);
+    if (borrar[1]) {
+      const archivosBorrar = [...borrar[1]];
 
-        //borro el fichero del servidor
-        let path_file = await path.join(
-          __dirname,
-          `..`,
-          `..`,
-          `discoDuro`,
-          `${idUsuario}`,
-          await ficheroBorrado
-        );
-        await fs.unlink(path_file);
+      if (archivosBorrar.length > 0 || archivosBorrar != undefined) {
+        //recorro el array de ficheros pertenecientes a la carpeta a borrar
+        for (const archivo of archivosBorrar) {
+          //borro el fichero en la base de datos
+          let ficheroBorrado = borrarFichero(idUsuario, archivo.id_archivo);
+
+          //borro el fichero del servidor
+          let path_file = await path.join(
+            __dirname,
+            `..`,
+            `..`,
+            `discoDuro`,
+            `${idUsuario}`,
+            await ficheroBorrado
+          );
+          await fs.unlink(path_file);
+        }
       }
     }
     //elimino de la base de datos la carpeta a eliminar
+
     const eliminado = await eliminarDirectorio(idUsuario, idCarpeta);
 
     res.send({
